Move loadPlainTextLog out into own source file
